Moved constants to sdk/ to remove spurious 'controlplane' dependencies.
