+++ k8s/cmd/nsmdp/nsmdp.go
@@ -31,6 +31,7 @@ import (
 	"github.com/networkservicemesh/networkservicemesh/controlplane/pkg/nsmd"
 	"github.com/networkservicemesh/networkservicemesh/controlplane/pkg/serviceregistry"
 	"github.com/networkservicemesh/networkservicemesh/pkg/tools"
+	"github.com/networkservicemesh/networkservicemesh/sdk/common"
 )
 
 const (
@@ -74,10 +75,10 @@ func (n *nsmClientEndpoints) Allocate(ctx context.Context, reqs *pluginapi.Alloc
 			responses.ContainerResponses = append(responses.ContainerResponses, &pluginapi.ContainerAllocateResponse{
 				Mounts: []*pluginapi.Mount{mount},
 				Envs: map[string]string{
-					nsmd.NsmDevicePluginEnv: "true",
-					nsmd.NsmServerSocketEnv: mount.ContainerPath + workspace.NsmServerSocket,
-					nsmd.NsmClientSocketEnv: mount.ContainerPath + workspace.NsmClientSocket,
-					nsmd.WorkspaceEnv:       workspace.ClientBaseDir,
+					nsmd.NsmDevicePluginEnv:   "true",
+					common.NsmServerSocketEnv: mount.ContainerPath + workspace.NsmServerSocket,
+					common.NsmClientSocketEnv: mount.ContainerPath + workspace.NsmClientSocket,
+					common.WorkspaceEnv:       workspace.ClientBaseDir,
 				},
 			})
 		}
